+++ packages/itwin/measure-tools/src/widgets/MeasurementActionToolbar.tsx
@@ -8,7 +8,7 @@ import type { XAndY } from "@itwin/core-geometry";
 import { Point2d } from "@itwin/core-geometry";
 import { RelativePosition } from "@itwin/appui-abstract";
 import type { ItemProps, ToolbarActionItem } from "@itwin/appui-react";
-import { ActionButtonItemDef, CursorInformation, CursorPopupManager, ToolbarItemUtilities, ToolbarOrientation, ToolbarUsage } from "@itwin/appui-react";
+import { ActionButtonItemDef, CursorInformation, CursorPopupManager, ToolbarItemUtilities } from "@itwin/appui-react";
 import { FeatureTracking, MeasureToolsFeatures } from "../api/FeatureTracking";
 import type { Measurement, MeasurementPickContext } from "../api/Measurement";
 import { MeasurementManager } from "../api/MeasurementManager";
@@ -442,11 +442,7 @@ export class MeasurementActionToolbar {
       itemDef.measurements = measurements;
       return ToolbarItemUtilities.createActionItem(itemDef.id, index * 10, itemDef.iconSpec, itemDef.label, itemDef.execute);
     });
-    return <PopupToolbar
-      items={toolItems}
-      usage={ToolbarUsage.ContentManipulation}
-      orientation={ToolbarOrientation.Horizontal}
-    />;
+    return <PopupToolbar items={toolItems} onClose={() => MeasurementActionToolbar.closeToolbar(true)}/>;
   }
 }