+++ packages/itwin/ec3-widget/CHANGELOG.md
@@ -1,6 +1,32 @@
 # Change Log - @itwin/ec3-widget-react
 
-This log was last generated on Tue, 18 Jul 2023 14:40:24 GMT and should not be manually modified.
+This log was last generated on Thu, 24 Aug 2023 21:32:34 GMT and should not be manually modified.
+
+## 0.4.1
+Thu, 24 Aug 2023 21:32:34 GMT
+
+### Patches
+
+- Modified toolbar layout in EC3 widget to be consistent with G&M widget.
+- Truncated long assembly name when in narrow width.
+
+## 0.4.0
+Fri, 18 Aug 2023 11:01:29 GMT
+
+### Minor changes
+
+- Migrated UI items for iTwin UI 2.x and iTwinJS 4.x upgrade.
+
+### Patches
+
+- Removed frame around Add Assembly button and changed button type.
+
+## 0.3.0
+Thu, 17 Aug 2023 14:42:37 GMT
+
+### Minor changes
+
+- Typescript version bump to ^4.5.0
 
 ## 0.2.0
 Tue, 18 Jul 2023 14:40:24 GMT